Change logs to be global
