+++ README.md
@@ -19,6 +19,8 @@ A set of plugins to interact with Datadog directly from your builds.
 - [Packages](#packages)
 - [Configuration](#configuration)
   - [`auth.apiKey`](#authapikey)
+  - [`auth.endPoint`](#authendpoint)
+  - [`logLevel`](#loglevel)
 - [Contributing](#contributing)
 - [License](#license)
 
@@ -34,7 +36,9 @@ A set of plugins to interact with Datadog directly from your builds.
 {
     auth: {
         apiKey: '<mydatadogkey>',
+        endPoint: 'app.datadoghq.com',
     },
+    logLevel: 'debug',
     [plugin-name]: {
         disabled: true,
         [...plugin-specific-configuration],
@@ -44,11 +48,21 @@ A set of plugins to interact with Datadog directly from your builds.
 
 ### `auth.apiKey`
 
-> required
+> default null
 
 In order to interact with Datadog, you have to use [your own API Key](https://app.datadoghq.com/account/settings#api).
 
-Without a key, the plugin won't send anything to Datadog.
+### `auth.endPoint`
+
+> default: `"app.datadoghq.com"`
+
+To which endpoint will the metrics be sent.
+
+### `logLevel`
+
+> default: 'warn'
+
+Which level of log do you want to show.
 
 ---
