+++ packages/plugins/telemetry/src/common/output/metrics.ts
@@ -3,17 +3,18 @@
 // Copyright 2019-Present Datadog, Inc.
 
 import { formatDuration } from '@dd/core/helpers';
+import { getLogFn } from '@dd/core/log';
 import c from 'chalk';
 
-import { CONFIG_KEY, PLUGIN_NAME } from '../../constants';
+import { PLUGIN_NAME } from '../../constants';
 import type { Context, OptionsWithTelemetryEnabled } from '../../types';
 import { getMetrics } from '../aggregator';
-import { getLogFn, getLogLevel, getMetric, getOptionsDD } from '../helpers';
+import { getMetric, getOptionsDD } from '../helpers';
 import { sendMetrics } from '../sender';
 
 export const addMetrics = (context: Context, options: OptionsWithTelemetryEnabled) => {
     const { report, bundler } = context;
-    const log = getLogFn(getLogLevel(options[CONFIG_KEY].output));
+    const log = getLogFn(options.logLevel, PLUGIN_NAME);
 
     context.metrics = context.metrics || [];
     try {
@@ -28,7 +29,7 @@ export const processMetrics = async (context: Context, options: OptionsWithTelem
     const { start } = context;
     const duration = Date.now() - start;
     const optionsDD = getOptionsDD(options);
-    const log = getLogFn(optionsDD.logLevel);
+    const log = getLogFn(options.logLevel, PLUGIN_NAME);
     context.metrics = context.metrics || [];
     // We're missing the duration of this hook for our plugin.
     context.metrics.push(
@@ -46,13 +47,13 @@ export const processMetrics = async (context: Context, options: OptionsWithTelem
     log(`Took ${formatDuration(duration)}.`);
 
     // Send everything only if we have the key.
-    if (!optionsDD.apiKey) {
+    if (!options.auth?.apiKey) {
         log(`Won't send metrics to ${c.bold('Datadog')}: missing API Key.`, 'warn');
         return;
     }
     try {
         const startSending = Date.now();
-        await sendMetrics(context.metrics, optionsDD);
+        await sendMetrics(context.metrics, options);
         log(`Sent metrics in ${formatDuration(Date.now() - startSending)}.`);
     } catch (e) {
         log(`Error sending metrics ${e}`, 'error');
