+++ packages/plugins/telemetry/src/common/sender.ts
@@ -2,18 +2,24 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
+import { getLogFn } from '@dd/core/log';
 import { request } from 'https';
 import type { ServerResponse } from 'http';
 
-import type { MetricToSend, OptionsDD } from '../types';
+import { PLUGIN_NAME } from '../constants';
+import type { MetricToSend, OptionsWithTelemetryEnabled } from '../types';
 
-import { getLogFn } from './helpers';
+export const sendMetrics = (metrics: MetricToSend[], opts: OptionsWithTelemetryEnabled) => {
+    const log = getLogFn(opts.logLevel, PLUGIN_NAME);
 
-export const sendMetrics = (metrics: MetricToSend[], opts: OptionsDD) => {
+    if (!opts.auth?.apiKey) {
+        log(`Won't send metrics to Datadog: missing API Key.`, 'warn');
+        return;
+    }
+
     if (!metrics || !metrics.length) {
         throw new Error('No metrics to send.');
     }
-    const log = getLogFn(opts.logLevel);
 
     const metricsNames = [...new Set(metrics.map((m) => m.metric))]
         .sort()
@@ -28,8 +34,8 @@ Metrics:
     return new Promise((resolve, reject) => {
         const req = request({
             method: 'POST',
-            hostname: opts.endPoint,
-            path: `/api/v1/series?api_key=${opts.apiKey}`,
+            hostname: opts.auth?.endPoint || 'app.datadoghq.com',
+            path: `/api/v1/series?api_key=${opts.auth?.apiKey}`,
         });
 
         req.write(
