+++ packages/plugins/telemetry/src/types.ts
@@ -7,8 +7,6 @@ import type { BuildOptions, BuildResult, Metafile } from 'esbuild';
 
 import type { CONFIG_KEY } from './constants';
 
-export type LogLevel = 'debug' | 'warn' | 'error' | 'none';
-
 export interface MetricToSend {
     type: 'gauge';
     tags: string[];
@@ -17,13 +15,10 @@ export interface MetricToSend {
 }
 
 export interface OptionsDD {
-    apiKey: string;
     tags: string[];
-    endPoint: string;
     prefix: string;
     timestamp: number;
     filters: Filter[];
-    logLevel: LogLevel;
 }
 
 export interface Metric {
@@ -37,7 +32,6 @@ export type Filter = (metric: Metric) => Metric | null;
 
 export type OutputOptions =
     | boolean
-    | LogLevel
     | string
     | {
           destination: string;
@@ -45,13 +39,11 @@ export type OutputOptions =
           dependencies?: boolean;
           bundler?: boolean;
           metrics?: boolean;
-          level?: LogLevel;
       };
 
 export type TelemetryOptions = {
     disabled?: boolean;
     output?: OutputOptions;
-    endPoint?: string;
     prefix?: string;
     tags?: string[];
     timestamp?: number;
